+++ spec/factories/users.rb
@@ -12,6 +12,9 @@ FactoryBot.define do
     trait :admin do
       admin { true }
     end
+
+    activated { true }
+    activated_at { Time.zone.now }
   end
 
   factory :user_update_after, class: User do
@@ -25,6 +28,9 @@ FactoryBot.define do
     trait :admin do
       admin { true }
     end
+
+    activated { true }
+    activated_at { Time.zone.now }
   end
 
   factory :other_user, class: User do
@@ -32,6 +38,8 @@ FactoryBot.define do
     sequence(:email) { |n| "example_other#{n}@example.com" }
     password { 'foobar' }
     password_confirmation { 'foobar' }
+    activated true
+    activated_at { Time.zone.now }
   end
 
   factory :invalid_user, class: User do
@@ -40,5 +48,18 @@ FactoryBot.define do
     # image { Rack::Test::UploadedFile.new(File.join(Rails.root, 'spec/fixtures/test.jpg')) }
     password { 'foobar' }
     password_confirmation { 'foobaz' }
+    activated false
+    activated_at { Time.zone.now }
+  end
+
+  factory :show_nil_user, class: User do
+    id { '900' }
+    name { nil }
+    sequence(:email) { nil }
+    # image { Rack::Test::UploadedFile.new(File.join(Rails.root, 'spec/fixtures/test.jpg')) }
+    password { nil }
+    password_confirmation { nil }
+    activated false
+    activated_at { Time.zone.now }
   end
 end
