+++ spec/fixtures/user_mailer/account_activation
@@ -1,0 +1,3 @@
+UserMailer#account_activation
+
+Hi, find me in app/views/user_mailer/account_activation
