+++ spec/fixtures/user_mailer/password_reset
@@ -1,0 +1,3 @@
+UserMailer#password_reset
+
+Hi, find me in app/views/user_mailer/password_reset
