+++ spec/helpers/account_activations_helper_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+# Specs in this file have access to a helper object that includes
+# the AccountActivationsHelper. For example:
+#
+# describe AccountActivationsHelper do
+#   describe "string concat" do
+#     it "concats two strings with spaces" do
+#       expect(helper.concat_strings("this","that")).to eq("this that")
+#     end
+#   end
+# end
+RSpec.describe AccountActivationsHelper, type: :helper do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
