+++ spec/mailers/user_mailer_spec.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe UserMailer, type: :mailer do
+  describe 'account_activation' do
+    let(:user) { FactoryBot.create(:user) }
+    let(:mail) { UserMailer.account_activation(user) }
+
+    it 'renders the headers' do
+      expect(mail.subject).to eq('Account activation')
+      expect(mail.to).to eq([user.email])
+      expect(mail.from).to eq(['noreply@example.com'])
+    end
+
+    it 'renders the body' do
+      expect(mail.body.encoded).to match user.name
+      expect(mail.body.encoded).to match user.activation_token
+      expect(mail.body.encoded).to match CGI.escape(user.email)
+    end
+  end
+
+  # describe 'password_reset' do
+  #   let(:mail) { UserMailer.password_reset }
+
+  #   it 'renders the headers' do
+  #     expect(mail.subject).to eq('Password reset')
+  #     expect(mail.to).to eq(['to@example.org'])
+  #     expect(mail.from).to eq(['from@example.com'])
+  #   end
+
+  #   it 'renders the body' do
+  #     expect(mail.body.encoded).to match('Hi')
+  #   end
+  # end
+end
