+++ spec/models/user_spec.rb
@@ -110,8 +110,8 @@ RSpec.describe User, type: :model do
     expect(user.errors[:password_confirmation]).to include("doesn't match Password")
   end
 
-  # remember_digestが存在しない場合のテスト
-  it 'return false for an user with nil digest' do
-    expect(user.authenticated?('')).to be_falsy
+  # authenticated?のテスト
+  it 'returns false for a user with nil digest' do
+    expect(user.authenticated?(:remember, '')).to be_falsy
   end
 end
